refactor(cli): extract shared helpers to app/cli/_shared.py
Moves console, _parse_inputs, _resolve_flow_id, _parse_iso,
_format_duration_seconds, _flow_duration_seconds, _format_clock,
_format_next_fire, _schedule_store out of app/main.py into the new
app/cli package. app/main.py re-imports them so command bodies that
still live there work unchanged.

Step 1 of the app/cli/ extraction. Tests stay green; --help output
remains byte-identical to tasks/baseline_help/.
